backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# File handling
python-multipart==0.0.9
aiofiles==23.2.1

# Image processing - Use pre-built wheel
Pillow>=10.0.0

# Environment variables
python-dotenv==1.0.1

# HTTP client
httpx==0.27.0

# QR Code generation
qrcode[pil]==7.4.2

# Additional dependencies for production
gunicorn==21.2.0
